+++ test/libsolidity/syntaxTests/dataLocations/transient_function_type.sol
@@ -1,0 +1,9 @@
+contract C {
+    function () transient f;
+    function (uint) external transient y;
+    function () transient internal fti;
+    function () internal transient fit;
+    function () internal transient internal fiti;
+    function () internal internal transient fiit;
+}
+// ----