+++ test/libsolidity/syntaxTests/dataLocations/transient_state_variable_visibility.sol
@@ -1,0 +1,10 @@
+contract C {
+    uint public transient pubt;
+    uint internal transient it;
+    uint private transient prvt;
+
+    uint transient public tpub;
+    uint transient internal ti;
+    uint transient private tprv;
+}
+// ----