+++ test/libsolidity/syntaxTests/dataLocations/transient_value_type_state_variables.sol
@@ -1,0 +1,10 @@
+contract D { }
+
+contract C {
+	address transient a;
+	bool transient b;
+	D transient d;
+	uint transient x;
+	bytes32 transient y;
+}
+// ----