+++ test/libsolidity/syntaxTests/dataLocations/transient_value_type_state_variables_assignment.sol
@@ -1,0 +1,8 @@
+contract D { }
+
+contract C {
+	int transient x = -99;
+	address transient a = address(0xABC);
+	bool transient b = x > 0 ? false : true;
+}
+// ----