+++ test/libsolidity/syntaxTests/dataLocations/variable_declaration_location_transient_test_reference_type.sol
@@ -1,0 +1,14 @@
+contract test {
+    struct S {
+        uint x;
+    }
+    S transient b;
+
+    uint[] transient a;
+    function f() public {
+        uint[] transient c;
+        S transient d;
+    }
+}
+// ----
+// ParserError 2314: (148-149): Expected ';' but got identifier