+++ test/libsolidity/syntaxTests/operators/transient_value_type.sol
@@ -1,0 +1,9 @@
+contract C {
+    int transient x;
+    function f() public view returns (int) {
+        int y = x;
+        int w = -x;
+        return (x + w) * (y / x);
+    }
+}
+// ----