+++ test/libsolidity/syntaxTests/types/address/state_variable_address_payable_transient.sol
@@ -1,0 +1,4 @@
+contract C {
+    address payable transient a;
+}
+// ----