+++ test/libsolidity/syntaxTests/variableDeclaration/transient_as_identifier_and_data_location.sol
@@ -1,0 +1,4 @@
+contract C {
+	uint transient transient;
+}
+// ----