+++ Changelog.md
@@ -17,6 +17,7 @@ Bugfixes:
 
 Language Features:
  * Introduce a new overload ``require(bool, Error)`` that allows usage of ``require`` functions with custom errors. This feature is available in the ``via-ir`` pipeline only.
+ * Accept declarations of state variables with ``transient`` data location (parser support only, no code generation yet).
 
 
 Compiler Features: